ui test: introduce SettingsWorkingTime page
