+++ src/main/resources/templates/settings/public-holidays/settings_working_time.html
@@ -43,7 +43,7 @@
             </div>
             <div class="form-section tw-mb-8">
               <div th:replace="~{settings/public-holidays/overtime::section}"></div>
-              <th:block th:replace="~{settings/fragments::save-button}" />
+              <th:block th:replace="~{settings/fragments::save-button(testId='submit-overtime-button')}" />
             </div>
           </div>
         </form>
